--- eBankMachine/card_update.h
@@ -1,7 +1,0 @@
-// ============================
-// FILE: card_update.h
-// ============================
-#pragma once
-void startCardUpdateFlow();
-void cardTick();
-void handleCardKey(char k);
