+++ eBankMachine/config.cpp
@@ -1,7 +1,7 @@
 // ============================
 // FILE: config.cpp
 // ============================
-#include "config.h"
+#include "eBankMachine.h"
 
 // ===== WiFi + Formbar =====
 const char* WIFI_SSID = "robonet";
