--- eBankMachine/deposit.h
@@ -1,7 +1,0 @@
-// ============================
-// FILE: deposit.h
-// ============================
-#pragma once
-void startDepositFlow();
-void depositTick();
-void handleDepositKey(char k);
