--- eBankMachine/drop.h
@@ -1,8 +1,0 @@
-// ============================
-// FILE: drop.h
-// ============================
-#pragma once
-
-void startDrop(int count);
-void finishDrop(const char* why);
-void dropTick(); // drop IR counting while dropping
