--- eBankMachine/keypad_router.h
@@ -1,5 +1,0 @@
-// ============================
-// FILE: keypad_router.h
-// ============================
-#pragma once
-void keypadTick();
