--- eBankMachine/net.h
@@ -1,5 +1,0 @@
-// ============================
-// FILE: net.h
-// ============================
-#pragma once
-void wifiEnsureConnected();
