--- eBankMachine/refund_tick.h
@@ -1,5 +1,0 @@
-// ============================
-// FILE: refund_tick.h
-// ============================
-#pragma once
-void refundTick();
