--- eBankMachine/withdraw.h
@@ -1,6 +1,0 @@
-// ============================
-// FILE: withdraw.h
-// ============================
-#pragma once
-void startWithdrawWizard();
-void handleWithdrawKey(char k);
